Possibility to disable library with property. Refactor of tests. Fixes in autoconfiguration.
