+++ example/src/main/java/io/github/wpik/extramdc/example/Application.java
@@ -2,9 +2,7 @@ package io.github.wpik.extramdc.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
 
-@EnableAspectJAutoProxy
 @SpringBootApplication
 public class Application {
     public static void main(String[] args) {
