+++ spring-boot-starter/src/test/java/io/github/wpik/extramdc/spring/EnabledWithPropertyTest.java
@@ -1,0 +1,9 @@
+package io.github.wpik.extramdc.spring;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
+
+@SpringBootTest
+@TestPropertySource(properties = {"extra-mdc.enabled=true"})
+public class EnabledWithPropertyTest extends EnabledTestBase {
+}
