+++ spring-boot-starter/src/test/java/io/github/wpik/extramdc/spring/EnabledWithoutPropertyTest.java
@@ -1,0 +1,8 @@
+package io.github.wpik.extramdc.spring;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class EnabledWithoutPropertyTest extends EnabledTestBase {
+}
+
